Database set up
What this code does now is sets up a database (only set up locally to my computer at the moment). What this code does is upon clicking one of the buttons, code in getData.py is ran. The first thing that is checked is if there is already an entry for that symbol in the database, if not its going to create one, if there is it is going to check when the last time it was updated was. If the data was updated within one week, nothing will happen. If the data has not been updated within the last week, it will be updated.