Updates to database and addition of time series functionality.
Moved data directory for database into the project file. This is an attempt to make the data available to everyone. Will likely take work on the end of each individual to get it to work properly. Working on setting up a time series database and API call. Currently is only getting 100 entries from API calls that should have over 5,000. Working on figuring out why this is.