Backend Functionality
Moved table classes to newly created models.py. Created three new tables. Created new functionalities that take the data stored in TimeSeriesDailyData and calculate weekly, monthly, and yearly averages for all numerical fields. These functions are only called when a new entry is added to the TimeSeriesDailyData table. So they will be ran only on the first data pull and when any new data is added.